+++ scripts/checkState.ts
@@ -5,8 +5,10 @@ import { NetworkProvider } from '@ton/blueprint';
 const HW_ID = 0x1337133713371337n;
 const FW_HASH = 0xdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefn;
 
+const CONTRACT_KQ = "kQBVqAhPv_ANWm0hfjJdLnQmvvC8_rQ_NEryVX3uFOUF05OP";
+
 export async function run(provider: NetworkProvider) {
-    const contractAddress = Address.parse('kQA2fMBzpJ8yOUtSTj8HAB2q1U37uDRoHNBRFqPbGFaBLvDO');
+    const contractAddress = Address.parse(CONTRACT_KQ);
     const tonSha = provider.open(TonSha.fromAddress(contractAddress));
 
     const isAuth = await tonSha.getIsAuthorized(HW_ID);